+++ tests/testthat/test-adorn-ns.R
@@ -194,3 +194,13 @@ test_that("no message thrown on grouped df input", {
                   adorn_percentages() %>%
                   adorn_ns())
 })
+
+test_that("adorn_ns works on single column data.frame with custom Ns if tidyselect is used, #456", {
+  adorned_single <- mtcars %>%
+    tabyl(am, cyl) %>%
+    adorn_percentages()
+  adorned_single <- adorned_single %>%
+    select(a = `4`) %>%
+    adorn_ns(ns = select(attr(adorned_single, "core"), a = `4`),,,, a)
+  expect_equal(stringr::str_sub(adorned_single$a, -4, -1), c(" (3)", " (8)"))
+})